requirements.txt
numpy
matplotlib
scipy

bitarray
crccheck
pyzmq